% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakAnno.R
\docType{data}
\name{peakAnno}
\alias{peakAnno}
\title{peakAnno: a package annotating peaks from m6A-seq and ChIP-seq/ATAC-seq/Cut&Tag.}
\format{
An object of class \code{function} of length 1.
}
\usage{
peakAnno(GTF,organism,up,down,peaktype,bedfile,outpath,outfile).
}
\arguments{
\item{GTF}{a GTF file used for peakAnno.}

\item{organism}{peaks called from which species.}

\item{up}{the start position of gene promoter used for peakAnno.}

\item{down}{the end position of gene promoter used for peakAnno.}

\item{peaktype}{peaks must include chr,start,end, and strand (+/-/.) used for peakAnno.}

\item{bedfile}{a peak file used for peakAnno.}

\item{outpath}{the path saving the output file.}

\item{outfile}{the file name for the output file.}
}
\value{
A peak annotation file will be returned at default.

One file will be returned at default.
}
\description{
This function peakAnno() in the peakAnno package.
}
\details{
peakAnno annotate peaks and return the genomic feature associated with peaks.
}
\note{
GTF files are available https://www.gencodegenes.org/.

up: upstream of TSS; default=2000.

down: downstream of TSS; default=0.

peaktype: default=m6A.

outpath: pathway of output file.

outfile: output file name.
}
\examples{
# run the function
# peakAnno(GTF,organism,up,down,peaktype,bedfile,outpath,outfile)
}
\author{
Benxia Hu
}
\keyword{datasets}
